cactus-1098: separate dependent fiend from field & small updates
