+++ docs/Forms/Getting Started.story.mdx
@@ -16,7 +16,7 @@ Installation is simple enough, just add the latest version of "@repay/cactus-for
 
 Cactus Form also has an optional setup step using [patch-package](https://www.npmjs.com/package/patch-package), described in more depth [below](#patching-react-final-form).
 
-Note that Cactus Form also exports final-form and react-final-form features that have not been edited to keep everything in one place. It should cover the vast majority of use cases but "final-form", "react-final-form" can still be imported separately for the original version of `Field`, `Form`, etc 
+Note that Cactus Form also exports final-form and react-final-form features that have not been edited to keep everything in one place. It should cover the vast majority of use cases but "final-form", "react-final-form" can still be imported separately for the original version of `Field`, `Form`, or if there's an uncommon feature that you need.
 
 ## Using React Final Form
