+++ modules/cactus-form/src/stories/DependentField.story.mdx
@@ -1,0 +1,54 @@
+import { Meta } from '@storybook/addon-docs'
+
+<Meta title="Cactus Form/Components/DependentField" />
+
+
+# DependentField
+
+`DependentField` is an extenstion of `Field`.
+It creates a normal `Field`, then uses Final Form's `registerField` API to subscribe
+to changes to the dependency fields.
+The extra props look like this:
+
+```
+interface DependentFieldProps extends FieldProps {
+  dependsOn: string | string[] | Record<string, DependencyConfig>;
+  onDependencyChange?: (state: FieldState, props: FieldProps) => void
+}
+
+interface DependencyConfig {
+  // The basic field subscription format, e.g. `value: true`
+  [K in keyof FieldSubscription]?: boolean
+  onChange?: (state: FieldState, props: FieldProps) => void
+}
+```
+
+It's important to note that `dependsOn` is used in a `useEffect` hook to create the field subscriptions,
+so to prevent unnecessary subscription turnover you should either use a constant or something like `useMemo`;
+that caveat does _not_ apply to `onDependencyChange`, however.
+
+If you pass the dependencies as a string or an array of field names, they will only subscribe
+to `value` changes and will all use the same change handler;
+you can still tell which field changed (for arrays) by looking at `state.name`.
+The second dependency format allows more flexibility in the subscription and change handling.
+
+The first argument to the change handler is the state of the field being watched;
+it's the same as the state passed to the callback for `registerField`.
+The second argument is the merged props of the `DependentField`; chances are the only
+one you'll need is `onChange`, which you can use to update the DependentField's
+value as the dependencies change.
+
+```
+const MyForm = (props) => (
+  <Form {...props}>
+    <Field name="power" type="number" />
+    <DependentField name="reaction" dependsOn="power" onDependencyChange={
+      (state, { onChange }) => {
+        onChange(state.value > 9000 ? 'unbelievable' : 'still good')
+      }
+    } />
+  </Form>
+)
+```
+
+`DependentField` also has the `configureDefaults` and `withDefaults` functions that `Field` has.
