+++ modules/cactus-form/src/stories/FormSpy.story.tsx
@@ -1,7 +1,7 @@
 import { noop } from 'lodash'
 import React from 'react'
 
-import { Field, Form, FormSpy } from '../'
+import { Field, Form, FormSpy as _FormSpy } from '../'
 import docsMeta from './FormSpy.story.mdx'
 
 delete docsMeta.includeStories
@@ -10,12 +10,12 @@ export default {
   ...docsMeta,
 } as const
 
-export const BasicUsage = () => (
+export const FormSpy = () => (
   <Form onSubmit={noop}>
     <Field name="field 1" label="Field 1" />
     <Field name="field 2" label="Field 2" />
-    <FormSpy subscription={{ dirty: true }}>
+    <_FormSpy subscription={{ dirty: true }}>
       {({ dirty }) => <div>the form is {dirty ? 'dirty' : 'clean'}</div>}
-    </FormSpy>
+    </_FormSpy>
   </Form>
 )
