+++ modules/cactus-form/src/stories/OtherTools.story.mdx
@@ -2,41 +2,41 @@ import { Meta } from '@storybook/addon-docs'
 
 <Meta title="Cactus Form/Components/Other Tools" />
 
-#  Other Cactus Form Exports
+# Other Cactus Form Exports
 The below are also exported from Cactus Form despite being unchanged from `final-form` and `react-final-form`. This is done so that everything can be imported through one library, instead of three.
 
-##  ARRAY_ERROR - Final Form
+### ARRAY_ERROR (Final Form)
 
-Example use [here](https://codesandbox.io/s/46ozq77v7?file=/index.js:159-170)
-Final Form Docs [here](https://final-form.org/docs/final-form/api#array_error)
+- Example use [here](https://codesandbox.io/s/46ozq77v7?file=/index.js:159-170)
+- Final Form Docs [here](https://final-form.org/docs/final-form/api#array_error)
 
-##  FORM_ERROR - Final Form
+### FORM_ERROR (Final Form)
 
-Example use [here](https://codesandbox.io/s/vql1q4kzv0?from-embed=&file=/index.js:254-528)
-Final Form Docs [here](https://final-form.org/docs/final-form/api#form_error)
+- Example use [here](https://codesandbox.io/s/vql1q4kzv0?from-embed=&file=/index.js:254-528)
+- Final Form Docs [here](https://final-form.org/docs/final-form/api#form_error)
 
-#  configOptions - Final Form
-Final Form Docs [here](https://final-form.org/docs/final-form/types/Config)
+### configOptions (Final Form)
+- Final Form Docs [here](https://final-form.org/docs/final-form/types/Config)
 
-#  createForm - Final Form
-Final Form Docs [here](https://final-form.org/docs/final-form/getting-started#usage)
+### createForm (Final Form)
+- Final Form Docs [here](https://final-form.org/docs/final-form/getting-started#usage)
 
-#  fieldSubscriptionItems - Final Form
+### fieldSubscriptionItems (Final Form)
 
-Final Form Docs [here](https://final-form.org/docs/final-form/api#fieldsubscriptionitems)
+- Final Form Docs [here](https://final-form.org/docs/final-form/api#fieldsubscriptionitems)
 
-#  formSubscriptionItems - Final Form
+### formSubscriptionItems (Final Form)
 
-Final Form Docs [here](https://final-form.org/docs/final-form/api#formsubscriptionitems)
+- Final Form Docs [here](https://final-form.org/docs/final-form/api#formsubscriptionitems)
 
-# useField - React Final Form
-Final Form Docs [here](https://final-form.org/docs/react-final-form/api/useField)
+### useField (React Final Form)
+- Final Form Docs [here](https://final-form.org/docs/react-final-form/api/useField)
 
-# useForm - React Final Form
-Final Form Docs [here](https://final-form.org/docs/react-final-form/api/useForm)
+### useForm (React Final Form)
+- Final Form Docs [here](https://final-form.org/docs/react-final-form/api/useForm)
 
-# useFormState - React Final Form
-Final Form Docs [here](https://final-form.org/docs/react-final-form/api/useFormState)
+### useFormState (React Final Form)
+- Final Form Docs [here](https://final-form.org/docs/react-final-form/api/useFormState)
 
-# withTypes - React Final Form
-This function returns an object with the Form and FormSpy Types
+### withTypes (React Final Form)
+- This function returns an object with the Form and FormSpy Types
